requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-jose
passlib[bcrypt]
